# presentaciones/cuentas-api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Cuentas","description":"Controlador de peticiones REST de Cuentas"}],"paths":{"/cuentas/{id}":{"get":{"tags":["Cuentas"],"operationId":"obtener","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Cuenta"}}}}}},"put":{"tags":["Cuentas"],"operationId":"actualizar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Cuenta"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Cuenta"}}}}}},"delete":{"tags":["Cuentas"],"operationId":"eliminar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/cuentas":{"get":{"tags":["Cuentas"],"operationId":"listar","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Cuenta"}}}}}}},"post":{"tags":["Cuentas"],"operationId":"crear","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Cuenta"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Cuenta"}}}}}}},"/cuentas/{id}/usuario":{"get":{"tags":["Cuentas"],"summary":"Obtiene el usuario asociado a la cuenta de bitcoins","operationId":"obtenerUsuario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Retorno el usuario con exito","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Usuario"}}}}}}}},"components":{"schemas":{"Cuenta":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"correo":{"type":"string"},"idUsuario":{"type":"integer","format":"int64"}}},"Usuario":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"password":{"type":"string"},"correo":{"type":"string"}}}}}}